br2: add fstab
TODO: this is a temporary fstab for /boot /root /var partitions on SDcard.
It is not suitable for A/B partitions scheme and needs to be updated.
Current idea is to keep read-only A/B partitions on eMMC and to
keep /var on SDCard. So fstab will be modified accordingly.

Signed-off-by: Sergey Matyukevich <geomatsi@gmail.com>